chore: :arrow_down: chore(package): update dependencies and fix eslint
Description:
This commit updates several dependencies in the package.json and package-lock.json files. It also includes a fix for eslint.

Details:
The updated dependencies include "lucide-react" and "@types/node". The version of "lucide-react" has been updated from "^0.379.0" to "^0.381.0", and "@types/node" has been updated from "20.12.12" to "20.13.0". The eslint version has been updated from "9.3.0" to "^8.54.0".

Changeset:
- Updated "lucide-react" version
- Updated "@types/node" version
- Updated eslint version

Modified files:
- package-lock.json
- package.json
This commit ensures that the project is using the latest versions of its dependencies and fixes eslint.

Signed-off-by: RicardoGEsteves(Ricardo Esteves) 30535937+RicardoGEsteves@users.noreply.github.com

Refs: [cd7855ade1150f3c8f863ed9844490ba44eb66f5]
Related to: [#213]